fix: fix panic in maintenance upgrade
Fix the regression introduced in https://github.com/siderolabs/omni/pull/1214.

It was wrong to use `MachineConfigGenOptions` in the new upgrade endpoint, as its `InstallImage` field is populated only if the machine was ever allocated.

So, for example, when using disk images, the machine is installed to disk (maintenance-upgradeable), but the `InstallImage` field is not populated.

Fix this by building the `InstallImage` spec using the `MachineStatus`.

Signed-off-by: Utku Ozdemir <utku.ozdemir@siderolabs.com>