fix: check config patch creation time as well as updated on orphan check
In the config patch cleanup controller, we falsely assumed that the resource update time was set to be the same as the creation time.

This in very rare cases caused the cleanup controller started by the runtime in the tests to prune the config patch created by the test itself, failing it sporadically.

This should fix the rare test flake we have:

```text
#66 4.431     grpc_test.go:266:
#66 4.431         	Error Trace:	/src/internal/backend/grpc/grpc_test.go:266
#66 4.431         	Error:      	Received unexpected error:
#66 4.431         	            	rpc error: code = NotFound desc = resource ConfigPatches.omni.sidero.dev(default/1@undefined) doesn't exist
#66 4.431         	Test:       	TestGrpcSuite/TestCrud
```

Signed-off-by: Utku Ozdemir <utku.ozdemir@siderolabs.com>