fix: make sure clipped req/resp content logs are still valid JSONs
When we clip the body JSON of a request or response because it is too large, the clipped bytes are not a valid JSON anymore, causing it to be logged like:
```
"grpc.request.content":{},"grpc.request.contentError":"json: error calling MarshalJSON for type *grpcutil.jsonpbObjectMarshaler: unexpected end of JSON input"
```

Try to address this by re-encoding the clipped JSON one more time, taking it as a raw string. This way we can still inspect the contents in the logs to some extent. An example:
```
"grpc.request.content": {"msg": "{\"publicKey\":{\"pgpData\":\"LS0tLS1CRUdJTiBQR1AgUFVCTElDIEtFWSBCTE9DSy0tLS0tCgp4ak1FYUZpZUFCW..."}
```

Signed-off-by: Utku Ozdemir <utku.ozdemir@siderolabs.com>