fix: correctly detect installation status for bare-metal machines
We read the block disks on the machine status if there is a system disk to detect if Talos is installed on the machine.
Based on this, we allow or disallow maintenance updates.

This value might be stale when the bare-metal provider is used, as it powers off the machines when they are idle, and this causes their hw information to not be updated. So, the disk might be wiped, but machine status poller might not catch the change in the block devices before it is powered off.

Instead, take the `installed` flag on the infra.MachineStatus resource as the source of truth for the machines managed by the bare metal infra provider. Update the `installed` label based on that.

Change the logic in the frontend for installation detection to use this label instead of checking block devices directly.

Signed-off-by: Utku Ozdemir <utku.ozdemir@siderolabs.com>