fix: fix the log spam caused by the expensive reqs to embedded etcd
The recent changes in the embedded etcd dependency caused it to log every request to it with the warning level due to the expensive request duration threshold being set to the zero value of duration. Fix this by setting the threshold to the default constant (300ms), which is the default for the etcd when it is run normally (not embedded).

Signed-off-by: Utku Ozdemir <utku.ozdemir@siderolabs.com>