fix: keep query parameters encoded in the oidc/saml login flows
We were setting the query parameters in the login flow incorrectly - we were reading the decoded query via `route.query` and setting them without re-encoding them.

This caused `redirect` query param, which is used in the workload proxy flow, to be double-decoded: once by the frontend and then by the server. The second decoding converted `+` to a ` ` (space), which caused base64 decoding to fail (`+` is a valid base64 char, while space is not).

Fix this by copying the raw query params as-is.

Fixes siderolabs/omni#1596.

Signed-off-by: Utku Ozdemir <utku.ozdemir@siderolabs.com>
(cherry picked from commit 99f1506f01339f581ef5ac265459071a2b9ffbc6)