fix: make workload proxy cookies HTTP only
It required moving cookies creation/deletion to the backend.
Cookies are now updated on `ConfirmPublicKey` and `RevokePublicKey` calls.

Introduced the `RevokePublicKey` method as `auth0` flow didn't have any
backend calls in the end of the logout.

Added a way for propagating cookies from the gRPC service to the gRPC
gateway.

Signed-off-by: Artem Chernyshev <artem.chernyshev@talos-systems.com>