feat: centralize Schematic ID computation
We compute the schematic id for a machine in two different places: in the `SchematicConfigurationController` for allocated machines, and in `MachineUpgradeStatusController` for maintenance mode machines.

Centralize this computation to be done only in `SchematicConfigurationController`.

Change the lifecycle of the `SchematicConfiguration` resource to be bound to a machine, not to a cluster.

Signed-off-by: Utku Ozdemir <utku.ozdemir@siderolabs.com>