fix: do not skip creating schematic config in agent mode
A recent change caused a regression in the bare metal provider, causing it to not attempt to power on the machines when they were allocated to a cluster.

The issue was caused by the following change, which resets the initial schematic state when the machine is detected to be in agent mode: https://github.com/siderolabs/omni/commit/15deddde56c560c1501e59ca4d7aa68743c49980#diff-6c417154d54297ede498c23ea4bd82a26982766196dbbb8e60aee26322830bcdR525

This caused the `SchematicConfigurationController` to skip reconciliation for the bare metal machines in agent mode, causing the `SchematicConfiguration` resources to not be created. This caused the `TalosVersion` to not be propagated to the `InfraMachine` resources, which caused the bare metal provider to not detect them as allocated.

Fix this in two places:
- Never reset the initial state to be `nil` in `MachineStatusController`.
- Check for the agent mode in `SchematicConfigurationController` and do not skip reconciliation if the machine is in agent mode. (Normally, this would not be needed, as the above change would be enough, but we have a special case here: `MachineStatus` will never be updated for powered off machines, so we cannot rely on it to bring us to the correct the state).

Signed-off-by: Utku Ozdemir <utku.ozdemir@siderolabs.com>