fix: prevent `MachineSetStatus` from going into create/destroy loop
The `MachineSetStatus` resource getting modified in the finalizer removal function was causing it to go into a create/delete loop, filling up etcd. This issue is only observable when the `MachineSet` had the `MachineProvisionController` finalizer (when an infra provider is used).

Change the modify function to effectively be an update, never create the resource.

The fix is tested in a staging environment and confirmed to be resolving the issue.

Co-authored-by: Artem Chernyshev <artem.chernyshev@talos-systems.com>
Signed-off-by: Utku Ozdemir <utku.ozdemir@siderolabs.com>