fix: reduce usage of cached state to avoid stale reads
Cached state usage should be very limited, as it has weaker consistency guarantees compared to the underlying state. Revert the usage of cached state in the validations and in various components, which we introduced a while ago.

The COSI resource server still uses the cached state.

Fix another cache-related stale read behavior in the `MachineUpgradeStatus` controller (same as 3e90bc6c94205d5b150766290558993d0ed208a6) and fix its test to not be flaky.

Signed-off-by: Utku Ozdemir <utku.ozdemir@siderolabs.com>