fix: move infra provider ID annotations to labels
This change updates the handling of the infra provider ID by migrating it from annotations to labels in the `Link` and `Machine` resources.

This was done because we have noticed that there was a bug in the `InfraMachineController`, where we mapped them to the primary input using a label query. It was also named as a label and was used as a label in other resource types.

Signed-off-by: Utku Ozdemir <utku.ozdemir@siderolabs.com>