fix: use deterministic order for machine extensions
When machine extensions are defined on multiple levels for a single machine, e.g., at cluster, machine set and cluster machine level, the one that was picked up as the actual configuration was determined randomly.

Add a unit test to reproduce this and fix it by calculating them using a deterministic order.

Co-authored-by: Oguz Kilcan <oguz.kilcan@siderolabs.com>
Signed-off-by: Utku Ozdemir <utku.ozdemir@siderolabs.com>