test: fix flaky `MachineUpgradeStatusController` test
Do the same thing we did in the schematic configuration controller and read the KernelArgs resource uncached.

This logic is in the process of being centralized in #1792, but still, it should help with the test stability at the time being.

Signed-off-by: Utku Ozdemir <utku.ozdemir@siderolabs.com>
(cherry picked from commit 4d0658bb106d6afce5385112452911abf04e5ae8)