fix: properly filter the machines which were already added to a cluster
The previous implementation of `ClusterMachineRequestStatus` watch was
incorrect: it was getting 25 first requests for clusters with many
machines, but as the sort order is different filtering on the frontend
side was missing some machines.

Move the filtering to the backend.
Now the requests which don't have `ClusterMachineStatuses` created will
not have `omni.sidero.dev/machine-request-in-use` label.
The UI now filters the requests which don't have this label to show them
while `ClusterMachineStatuses` doesn't exist.

Signed-off-by: Artem Chernyshev <artem.chernyshev@talos-systems.com>