feat: allow passing extra parameters to sqlite conn string
Do not expose the base parameters, as they are required for COSI to work correctly.

Add a flag for the additional parameters, which can for example be used for configuring the SQLite db to be "suitable" for litestream to be able to make backups of it.

Also, add `_pragma=synchronous(NORMAL)` to the base parameters, as it is recommended for better performance when WAL mode is enabled.

Signed-off-by: Utku Ozdemir <utku.ozdemir@siderolabs.com>